amend

+++ 0x13-more_singly_linked_lists/102-free_listint_safe.c
@@ -9,7 +9,7 @@ size_t free_listint_safe(listint_t **h)
 {
 	listint_t *temp;
 	size_t num = 0;
-	long int diff;
+	int diff;
 
 	if (!*h || !h)
 	{
